+++ includes/listeners/class-listenerfactory.php
@@ -45,7 +45,6 @@ class ListenerFactory {
 		'index.php',
 		'class-abstractlistener.php',
 		'class-listenerfactory.php',
-		'class-wordfencelistener.php',
 		'class-wpseolistener.php',
 	];
 
@@ -57,7 +56,7 @@ class ListenerFactory {
 	 */
 	private $late_init = [
 		'class-wsallistener.php',
-		//'class-wordfencelistener.php',
+		'class-wordfencelistener.php',
 	];
 
 	/**